Puts all tasks unit tests into subTest context (#500)
# Description

This MR moves all environment tasks names into `with.subTest()` context
manager to report per sub test as well.

## Type of change

- Bug fix (non-breaking change which fixes an issue)

## Checklist

- [x] I have run the [`pre-commit` checks](https://pre-commit.com/) with
`./orbit.sh --format`
- [ ] I have made corresponding changes to the documentation
- [x] My changes generate no new warnings
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [x] I have run all the tests with `./orbit.sh --test` and they pass
- [ ] I have updated the changelog and the corresponding version in the
extension's `config/extension.toml` file
- [x] I have added my name to the `CONTRIBUTORS.md` or my name already
exists there

---------

+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/utils/array.py
@@ -75,7 +75,7 @@ def convert_to_torch(
         tensor = array
     elif isinstance(array, np.ndarray):
         if array.dtype == np.uint32:
-            array = array.astype(np.int64)
+            array = array.astype(np.int32)
         # need to deal with object arrays (np.void) separately
         tensor = torch.from_numpy(array)
     elif isinstance(array, wp.array):
